requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
asttokens==3.0.0
attrs==25.3.0
backoff==2.2.1
beautifulsoup4==4.13.3
browser-use==0.1.40
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
Crawl4AI==0.5.0.post4
cryptography==44.0.2
cssselect==1.3.0
debugpy==1.8.13
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
executing==2.2.0
fake-http-header==0.3.5
fake-useragent==2.1.0
faust-cchardet==2.1.19
filelock==3.18.0
filetype==1.2.0
frozenlist==1.5.0
fsspec==2025.3.0
google-api-core==2.24.2
google-api-python-client==2.165.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-genai==1.7.0
googleapis-common-protos==1.69.2
greenlet==3.1.1
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.29.3
humanize==4.12.1
idna==3.10
importlib_metadata==8.6.1
iniconfig==2.1.0
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
langchain-anthropic==0.3.3
langchain-core==0.3.46
langchain-google-genai==2.1.0
langchain-ollama==0.2.2
langchain-openai==0.3.1
langsmith==0.3.18
litellm==1.63.12
lxml==5.3.1
markdown-it-py==3.0.0
markdownify==0.14.1
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
monotonic==1.6
multidict==6.2.0
nest-asyncio==1.6.0
nltk==3.9.1
numpy==2.2.4
ollama==0.4.7
openai==1.67.0
orjson==3.10.15
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.7
playwright==1.51.0
pluggy==1.5.0
posthog==3.21.0
prompt_toolkit==3.0.50
propcache==0.3.0
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pyee==12.1.1
Pygments==2.19.1
pyOpenSSL==25.0.0
pyparsing==3.2.1
pyperclip==1.9.0
pytest==8.3.5
pytest-base-url==2.1.0
pytest-playwright==0.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
PyYAML==6.0.2
pyzmq==26.3.0
rank-bm25==0.2.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.23.1
rsa==4.9
setuptools==77.0.1
six==1.17.0
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.6
stack-data==0.6.3
tenacity==9.0.0
text-unidecode==1.3
tf-playwright-stealth==1.1.2
tiktoken==0.9.0
tokenizers==0.21.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
wcwidth==0.2.13
websockets==15.0.1
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0